Excecao Delete e algumas outras alteracoes

+++ backend/src/main/java/com/lojinhasystem/system/services/CategoriaService.java
@@ -2,8 +2,10 @@ package com.lojinhasystem.system.services;
 
 import com.lojinhasystem.system.entities.Categoria;
 import com.lojinhasystem.system.repositories.CategoriaRepository;
+import com.lojinhasystem.system.services.exceptions.DatabaseException;
 import com.lojinhasystem.system.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -40,6 +42,14 @@ public class CategoriaService {
     }
 
     public void delete(Long id) {
-        categoriaRepository.deleteById(id);
+        try {
+            if (categoriaRepository.existsById(id)) {
+                categoriaRepository.deleteById(id);
+            } else {
+                throw new ResourceNotFoundException(id);
+            }
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 }
