+++ backend/src/main/java/com/lojinhasystem/system/services/ClienteService.java
@@ -2,8 +2,10 @@ package com.lojinhasystem.system.services;
 
 import com.lojinhasystem.system.entities.Cliente;
 import com.lojinhasystem.system.repositories.ClienteRepository;
+import com.lojinhasystem.system.services.exceptions.DatabaseException;
 import com.lojinhasystem.system.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -46,6 +48,14 @@ public class ClienteService {
     }
 
     public void delete(Long id) {
-        clienteRepository.deleteById(id);
+        try {
+            if (clienteRepository.existsById(id)) {
+                clienteRepository.deleteById(id);
+            } else {
+                throw new ResourceNotFoundException(id);
+            }
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 }
